Added a statisticsFunctions util file
