+++ scripts/comparing-staging-production.ts
@@ -5,6 +5,7 @@ import * as z from 'zod';
 import * as schemas from '../src/api/schemas';
 import { fetchCompanies } from './utils/fetchUtils';
 import { convertCompanyEvalsToCSV, generateXLSX } from './utils/outputFunctions';
+import { reportStatistics, outputTotalStatistics } from './utils/statisticsFunctions';
 
 type CompanyList = z.infer<typeof schemas.CompanyList>;
 type ReportingPeriod  = z.infer<typeof schemas.MinimalReportingPeriodSchema>;
@@ -12,7 +13,7 @@ type CompanyResponse = z.infer<typeof schemas.MinimalCompanyBase>;
 const NUMBER_OF_CATEGORIES = 16;
 const ONLY_CHECK_VERIFIED_DATA = true;
 
-interface Diff {
+export interface Diff {
   productionValue?: number;
   stagingValue?: number;
   difference?: number;
@@ -25,7 +26,7 @@ export interface Company {
   diffs: DiffReport[];
 }
 
-interface DiffReport {
+export interface DiffReport {
   reportingPeriod: {
     startDate: Date;
     endDate: Date;
@@ -137,7 +138,6 @@ function compareReportingPeriods(productionReportingPeriod: ReportingPeriod, sta
   );
   diffs.push(d.diffs.emissions.scope2.lb);
 
-
   d.diffs.emissions.scope2.mb = compareNumbers(
     productionReportingPeriod.emissions?.scope2?.mb,
     stagingReportingPeriod.emissions?.scope2?.mb,
@@ -215,54 +215,6 @@ function compareNumbers(productionNumber: number | undefined | null, stagingNumb
   return diff
 }
 
-function reportStatistics(diffs: Diff[]) {
-  const numbCorrectFieldsIncludeUndefined = diffs.reduce((acc: number, current: Diff) => {
-    return current.productionValue === current.stagingValue ? acc+1 : acc; // this also captures if both are undefined
-  }, 0);
-  // How many of the fields are correct?
-  const accuracy = {
-    description: 'Out of all fields, how many are correct?',
-    value: diffs.length > 0 ? numbCorrectFieldsIncludeUndefined / diffs.length : undefined,
-    numbCorrectFieldsIncludeUndefined,
-    numbFields: diffs.length
-  }
-  // Out of all fields that are supposed to have a numerical value, How many are correct? (excludes all instances where prod has an undefined value)
-  const numbCorrectNumericalFields = diffs.reduce((acc: number, current: Diff) => { return !(current.productionValue === undefined && current.stagingValue === undefined) && (current.productionValue === current.stagingValue) ? acc + 1 : acc;}, 0)
-  const numbNumericalFields = diffs.reduce((acc: number, current: Diff) => { return !(current.productionValue === undefined && current.stagingValue === undefined) ? acc + 1 : acc;}, 0);
-  const accuracyNumericalFields = {
-    description: 'Out of all fields that are supposed to have a numerical value, how many are correct?',
-    value: numbCorrectNumericalFields/numbNumericalFields,
-    numbCorrectNumericalFields,
-    numbNumericalFields
-  }
-  // Out of all fields that are supposed to have a numerical value, how many are incorrect because of a magnitude error?
-  const magErr = diffs.reduce((acc: number, current: Diff) => {
-    return (current.productionValue !== undefined) && 
-          (current.stagingValue !== undefined) && 
-          ((Math.log10(current.productionValue / current.stagingValue) % 1) === 0 ) ? 
-          acc + 1 : acc;
-  }, 0);
-  const magnError = magErr/numbNumericalFields
-
-  // // Out of all fields that are supposed to have a numerical value, how many have swapped fields?
-  // for(let i; i < diffs.length; i++) {
-  //   const prod = diffs[i].productionValue
-  //   for (let j; j < diffs.length; j++) {
-  //     if () {
-
-  //     }
-  //   }
-  // }
-  const fieldswap = diffs.reduce((acc: number, current: Diff) => { return current.productionValue !== undefined ? acc + 1 : acc;}, 0);
-
-  return {
-    accuracy,
-    accuracyNumericalFields,
-    magnError
-  }
-}
-
-
 async function outputEvalMetrics(companies: Company[]) {
   const outputPath = resolve('output', 'accuracy-results.csv');
   const outputXLSX = resolve('output', 'accuracy-results.xlsx');
@@ -271,37 +223,6 @@ async function outputEvalMetrics(companies: Company[]) {
   await writeFile(outputXLSX, xlsx, 'utf8');
   await writeFile(outputPath, csvContent, 'utf8');
   console.log(`✅ Statistics per report, written to ${outputPath}.`);
-}
-
-function outputTotalStatistics(companies: Company[]) {
-
-  const sumCorrectFieldsIncludeUndefined = companies.reduce((acc1: number, company: Company) => {
-    const sumCompanyCorrectFields = company.diffs.reduce((acc2: number, diff: DiffReport) => {
-      return diff.eval.accuracy?.numbCorrectFieldsIncludeUndefined ? acc2 + diff.eval.accuracy?.numbCorrectFieldsIncludeUndefined : acc2
-    }, 0)
-    return acc1 + sumCompanyCorrectFields
-  }, 0)
-  const sumNumbFields = companies.reduce((acc1: number, company: Company) => {
-    const sumCompanyFields = company.diffs.reduce((acc2: number, diff: DiffReport) => {
-      return diff.eval.accuracy?.numbFields ? acc2 + diff.eval.accuracy?.numbFields : acc2
-    }, 0)
-    return acc1 + sumCompanyFields
-  }, 0)
-
-  const sumCorrectFields = companies.reduce((acc1: number, company: Company) => {
-    const sumCompanyCorrectFields = company.diffs.reduce((acc2: number, diff: DiffReport) => {
-      return diff.eval.accuracyNumericalFields?.numbCorrectNumericalFields ? acc2 + diff.eval.accuracyNumericalFields?.numbCorrectNumericalFields : acc2
-    }, 0)
-    return acc1 + sumCompanyCorrectFields
-  }, 0)
-  const sumFields = companies.reduce((acc1: number, company: Company) => {
-    const sumCompanyFields = company.diffs.reduce((acc2: number, diff: DiffReport) => {
-      return diff.eval.accuracyNumericalFields?.numbNumericalFields ? acc2 + diff.eval.accuracyNumericalFields?.numbNumericalFields : acc2
-    }, 0)
-    return acc1 + sumCompanyFields
-  }, 0)
-  console.log(`sumCorrectFieldsIncludingUndefined: ${sumCorrectFieldsIncludeUndefined}, sumFieldsIncludingUndefined: ${sumNumbFields}`)
-  console.log(`sumCorrectFields: ${sumCorrectFields}, sumFields: ${sumFields}`)
 }
 
 // Main function for fetching, comparison, and output
